Make validating a captcha multiple times in 1 request possible by caching the result
git-svn-id: https://svn.muze.nl/svn/ariadne-php5/trunk@5020 7882d7ee-46e0-0310-a255-a28a61b628e4

+++ lib/modules/mod_captcha/hn_captcha.class.php
@@ -275,6 +275,8 @@ class hn_captcha
 		/** @private **/
 		var $b;
 
+		private static $validated = array();
+
 
 	////////////////////////////////
 	//
@@ -677,18 +679,25 @@ class hn_captcha
 		/** @private **/
 		function check_captcha($public,$private)
 		{
-			$res = 'FALSE';
-			// when check, destroy picture on disk
-			if(file_exists($this->get_filename($public)))
-			{
-				$res = @unlink($this->get_filename($public)) ? 'TRUE' : 'FALSE';
-				if($this->debug) echo "\n<br>-Captcha-Debug: Delete image (".$this->get_filename($public).") returns: ($res)";
-				$res = (strtolower($private)==strtolower($this->generate_private($public))) ? 'TRUE' : 'FALSE';
-				if($this->debug) echo "\n<br>-Captcha-Debug: Comparing public with private key returns: ($res)";
-			} else {
-				if($this->debug) echo "\n<br>-Captcha-Debug: Expired";
+			if(!isset(self::$validated[$private])) {
+				$res = 'FALSE';
+				// when check, destroy picture on disk
+				if(file_exists($this->get_filename($public)))
+				{
+					$res = @unlink($this->get_filename($public)) ? 'TRUE' : 'FALSE';
+					if($this->debug) echo "\n<br>-Captcha-Debug: Delete image (".$this->get_filename($public).") returns: ($res)";
+
+					$res = (strtolower($private)==strtolower($this->generate_private($public))) ? 'TRUE' : 'FALSE';
+					if($this->debug) echo "\n<br>-Captcha-Debug: Comparing public with private key returns: ($res)";
+
+				} else {
+					if($this->debug) echo "\n<br>-Captcha-Debug: Expired";
+				}
+				self::$validated[$private] = ($res == 'TRUE' ? TRUE : FALSE);
 			}
-			return $res == 'TRUE' ? TRUE : FALSE;
+
+			if($this->debug) echo "\n<br>-Captcha-Debug: returning validation result (".self::$validated[$private].")";
+			return self::$validated[$private];
 		}
 
 		/** @private **/
